Simplify simulations
Turns out it is possible to get GHDL to exit a simulation gracefully.
We just need the clock to stop and it GHDL will recognize that there are
no more events to process.

Also, we don't need to re-declare components. The code I was using as
inspiration was just bad I think.