Swap graphics mode to 800x600
I'm planning to use the 2 MiB SRAM as a pair of frame buffers, and
800x600x16bit is the ideal size.

This also adds the 100 MHz system clock. I don't know if that's a
realistic target for the system (especially the CPU), but we'll find
out.

Pixels will flow from the system clock domain into the VGA clock domain
via a FIFO.